Main UI is almost set up. ForEach Loop with list is set up
Need to set up:
- Summary View
- End of Day View
- Adding Date and Swiping Functionality
- Calculating the Carbon Footprints

+++ LogManager.swift
@@ -10,7 +10,7 @@ import Foundation
 struct CarbonLog: Identifiable, Codable {
     var id = UUID()
     var name: String
-    var footprint: Float
+    var footprint: Int
 }
 
 class CarbonLogManager: ObservableObject {
